schedule: consolidate ParseTimezone function into schedule package
There was a duplicated `ParseTimezone` function in both `materialize-sql` and
`sqlcapture` that was in use by a handful of connectors. This combines both of
them into a common location within the `schedule` package. The main reason for
doing this is that I will need to use the functionality of `ParseTimezone` in a
new type of Schedule that will be added in subsequent commits, and I didn't want
to have a third version of it.
